fix(i3): change paths in "include" jinja macro
The jinja macro used to simulate "include" in the i3 config file was
resolving paths in three different ways:
- use the path as-is
- relative to the current dotfile directory
- relative to i3-sway templating directory

Turns out that the first only makes sense in sway, as all i3 included
paths are within the repo and resolved *before* dotfiles installations.
Hence, it's been removed.

As for the other two, they both require a basename of the included path,
because the directory structure of those dotfiles in this repository is
flat.