refactor(debian): remove PATH bin linking
This was one of the earliest pieces of dotfiles I've ever written.
Still, I no longer use any of that software beside halt and reboot.
However, symlinking them and then setting SUID on the original file is
not compatible with the latest releases of systemd (the original file is
systemctl behind a layer of symlinks). Since there's a more proper way
to achieve the same result (userland passwordless poweroff and reboot),
the links for halt and reboot are no longer necessary and have been
removed.